requirements.txt
line-bot-sdk==1.8.0
flask==1.1.1
gunicorn
